src/requirements.txt
Pillow==9.1.0
python-dotenv==0.19.2